Migrate fjs/djs/ast to .f.mjs; extract fjs/djs's recursive types
fjs/djs/ast/module migrates from authored TypeScript to JSDoc-typed
JavaScript under the stage-1 migration. AstModule, AstConst,
AstModuleRef, AstArray, AstObject, and AstBody move to a sibling
types.ts since fjs/djs/parser and fjs/djs/transpiler depend on them.

This also extracts fjs/djs's own recursive Primitive/Object/Array/
Unknown types out of fjs/djs/module.f.ts into a new fjs/djs/types.ts,
hand-authored the same way fjs/media/json/types.ts's recursive Unknown
is (native TS type aliases, not JSDoc typedefs) — the pattern that
avoided the declaration-emit incompatibility JSDoc-typedef-declared
recursive types hit against rtti's own recursive Unknown during the
fjs/media/json migration. fjs/djs/module.f.ts re-exports the types and
stays TypeScript for now: it still depends on fjs/djs/serializer and
fjs/djs/transpiler, neither yet migrated. This unblocks ast, and will
unblock serializer and tokenizer next, without waiting for the whole
djs cluster to migrate at once.

Updates fjs/djs/ast/proof.f.ts, fjs/djs/parser/module.f.ts, and
fjs/djs/transpiler/module.f.ts import specifiers, and fixes stale
module.f.ts doc references (line-number citations split across the
new module.f.mjs/types.ts pair in fjs/djs/todo/*.md, a prose reference
in fjs/djs/README.md).
