+++ fjs/djs/ast/proof.f.ts
@@ -1,5 +1,5 @@
 import { sort } from '../../types/object/module.f.mjs'
-import { run } from './module.f.ts'
+import { run } from './module.f.mjs'
 import { stringifyAsTree } from '../serializer/module.f.ts'
 import { assertEq } from '../../asserts/module.f.mjs'
