+++ fjs/djs/parser/module.f.ts
@@ -12,7 +12,7 @@ import type { DjsToken, DjsTokenWithMetadata } from '../tokenizer/module.f.ts'
 import { setReplace, at } from '../../types/ordered_map/module.f.mjs'
 import type { OrderedMap } from '../../types/ordered_map/types.ts'
 import { fromMap } from '../../types/object/module.f.mjs'
-import type { AstArray, AstConst, AstModule, AstModuleRef } from '../ast/module.f.ts'
+import type { AstArray, AstConst, AstModule, AstModuleRef } from '../ast/types.ts'
 import type { TokenMetadata } from '../../js/tokenizer/types.ts'
 import { assertEq } from '../../asserts/module.f.mjs'
