+++ fjs/djs/todo/todjs-object-fold-is-a-map.md
@@ -5,14 +5,14 @@
 
 ### Problem
 
-In `fjs/djs/ast/module.f.ts` the two container branches of `toDjs` are
-asymmetric. The array branch is declarative (`:63`):
+In `fjs/djs/ast/module.f.mjs` the two container branches of `toDjs` are
+asymmetric. The array branch is declarative (`:54`):
 
 ```ts
 case 'array': { return toArray(map(toDjs(state))(ast[1])) }
 ```
 
-but the object branch (`:66-67`) threads a bespoke accumulator type through a
+but the object branch (`:57-58`) threads a bespoke accumulator type through a
 `fold`:
 
 ```ts
@@ -66,7 +66,8 @@ same `toDjs(state)` transformation; only the disguise is removed.
 
 ### Related
 
-- `fjs/djs/ast/module.f.ts:29-32`, `:41-46`, `:48-70` — the types and code
+- `fjs/djs/ast/types.ts:19-21`, `:30-36`, `:48-51` and
+  `fjs/djs/ast/module.f.mjs:18-25` — the types and code
   involved.
 - [197](./197.md) — the cross-function `Unknown`-walker factory; it lists
   `toDjs` as a caveat and defers its internals, so this intra-function
