+++ fjs/djs/transpiler/module.f.ts
@@ -14,7 +14,8 @@ import type { OrderedMap } from '../../types/ordered_map/types.ts'
 import { stringToList } from '../../text/utf16/module.f.mjs'
 import { concat as pathConcat } from '../../path/module.f.mjs'
 import { type ParseError, parseFromTokens } from '../parser/module.f.ts'
-import { run, type AstModule } from '../ast/module.f.ts'
+import { run } from '../ast/module.f.mjs'
+import type { AstModule } from '../ast/types.ts'
 import { foldStep, pure, step } from '../../effects/module.f.mjs'
 import type { Effect } from '../../effects/types.ts'
 import { readUtf8File } from '../../effects/node/module.f.mjs'
